Generated Core Scaffolds and Setup Initial Recipe Form
Scaffolds generated include ingredients, recipes, and recipe notes.
Did some Error Driven Development to remember how to get forms to work and created a preliminary recipe form with all the fields available. This will need to be styled and also integrated with ingredients once those are created.

Also set the root route to go to recipes index as this is where I think I will want to land for quick recipe look-up. I may also create a specific page for search later.

No tests yet this will also need to be covered. Still deciding whether I want to switch to rspec which I have previous familiarity with or challenge myself to use the test files generated as a learning experience. I will also use this as an opportunity to review Cucumber.